+++ modules/features/note/src/main/java/net/pantasystem/milktea/note/renote/RenotesViewModel.kt
@@ -13,13 +13,9 @@ import net.pantasystem.milktea.common.Logger
 import net.pantasystem.milktea.common.PageableState
 import net.pantasystem.milktea.common.StateContent
 import net.pantasystem.milktea.common.runCancellableCatching
-import net.pantasystem.milktea.model.notes.NoteRelationGetter
-import net.pantasystem.milktea.data.infrastructure.notes.renote.RenotesPagingService
-import net.pantasystem.milktea.model.notes.Note
-import net.pantasystem.milktea.model.notes.NoteCaptureAPIAdapter
-import net.pantasystem.milktea.model.notes.NoteRelation
-import net.pantasystem.milktea.model.notes.NoteRepository
+import net.pantasystem.milktea.model.notes.*
 import net.pantasystem.milktea.model.notes.renote.Renote
+import net.pantasystem.milktea.model.notes.renote.RenotesPagingService
 import net.pantasystem.milktea.model.user.User
 
 class RenotesViewModel @AssistedInject constructor(
