+++ modules/model/src/main/java/net/pantasystem/milktea/model/notes/renote/RenotesPagingService.kt
@@ -1,0 +1,17 @@
+package net.pantasystem.milktea.model.notes.renote
+
+import kotlinx.coroutines.flow.Flow
+import net.pantasystem.milktea.common.PageableState
+import net.pantasystem.milktea.model.notes.Note
+
+interface RenotesPagingService {
+
+    interface Factory {
+        fun create(noteId: Note.Id): RenotesPagingService
+    }
+
+    val state: Flow<PageableState<List<Renote>>>
+    suspend fun next()
+    suspend fun refresh()
+    suspend fun clear()
+}
